% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RadixS7.R
\name{erase}
\alias{erase}
\title{Erase sequences}
\usage{
erase(x, sequences)
}
\arguments{
\item{x}{A seqtrie S7 object.}

\item{sequences}{A character vector of sequences.}
}
\value{
A logical vector indicating whether each sequence was erased.
}
\description{
Erase sequences
}
