% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{generate_cost_matrix}
\alias{generate_cost_matrix}
\title{Generate a simple cost matrix}
\usage{
generate_cost_matrix(
  alphabet,
  match = 0,
  mismatch = 1,
  gap = 1,
  include_gap = FALSE
)
}
\arguments{
\item{alphabet}{A string representing all possible characters in both query and target sequences (e.g. "ACGT")}

\item{match}{The cost of a match}

\item{mismatch}{The cost of a mismatch}

\item{gap}{The cost of a gap, ignored if `include_gap` is FALSE (default: 1)}

\item{include_gap}{Whether to encode the cost of a gap in the matrix directly (default: FALSE)}
}
\value{
A cost matrix
}
\description{
Generate a cost matrix for use with the \code{search} method
}
\examples{
generate_cost_matrix("ACGT", match = 0, mismatch = 1)
}
