% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RadixS7.R
\name{has_sequence}
\alias{has_sequence}
\title{Test sequence membership}
\usage{
has_sequence(x, query)
}
\arguments{
\item{x}{A seqtrie S7 object.}

\item{query}{A character vector of query sequences.}
}
\value{
A logical vector indicating whether each query is present.
}
\description{
Test sequence membership
}
