% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RadixS7.R
\name{insert}
\alias{insert}
\title{Insert sequences}
\usage{
insert(x, sequences)
}
\arguments{
\item{x}{A seqtrie S7 object.}

\item{sequences}{A character vector of sequences.}
}
\value{
A logical vector indicating whether each sequence was inserted.
}
\description{
Insert sequences
}
