% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RadixS7.R
\name{is_valid}
\alias{is_valid}
\title{Validate trie or forest structure}
\usage{
is_valid(x)
}
\arguments{
\item{x}{A seqtrie S7 object.}
}
\value{
A logical value indicating whether the object is valid.
}
\description{
Validate trie or forest structure
}
