% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RadixS7.R
\name{plot_tree}
\alias{plot_tree}
\title{Plot trie or forest structure}
\usage{
plot_tree(x, depth = -1, root_label = "root", plot = TRUE)
}
\arguments{
\item{x}{A seqtrie S7 object.}

\item{depth}{The tree depth to plot. If -1, plot the entire tree.}

\item{root_label}{The label of the root node in the plot.}

\item{plot}{Whether to create a plot or return the graph data.}
}
\value{
A data frame of parent-child relationships or a ggplot2 object.
}
\description{
Plot trie or forest structure
}
