% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RadixS7.R
\name{prefix_search}
\alias{prefix_search}
\title{Prefix search}
\usage{
prefix_search(x, query)
}
\arguments{
\item{x}{A seqtrie S7 object.}

\item{query}{A character vector of query sequences.}
}
\value{
A data frame with columns \code{query} and \code{target}.
}
\description{
Prefix search
}
