% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zz_help_files.R
\name{PrefixTree}
\alias{PrefixTree}
\title{PrefixTree}
\description{
This class is a generic Prefix Tree class
}
\section{Usage}{

\preformatted{tree <- PrefixTree$new()

tree$print()

tree$to_string()

tree$to_dataframe()

tree$size()

tree$insert(sequences)

tree$erase(sequences)

tree$find(sequences)

tree$search(sequences, max_distance = NULL, max_fraction = NULL, mode = "levenshtein", nthreads = 1, display_progress = FALSE)
}
}

\section{Arguments}{

\describe{
  \item{sequences}{- The sequences to operate on.}
  \item{max_distance}{- For the search method, how far to search for similar sequences, in units of absolute distance. See details.}
  \item{max_fraction}{- For the search method, how far to search for similar sequences, relative to the sequence length. See details.}
  \item{mode}{- Either Levenshtein or Hamming. Levenshtein will allows for insertions and deletions and calculates "edit distance". Hamming does not allow for insertions or deletions.}
  \item{nthreads}{- How many threads to use in the search.}
  \item{display_progress}{- Display search progress (only for nthreads = 1).}
}
}

\section{Details}{

\code{$new()} creates a new Tree object, which holds the pointer to the underlying C++ implentation. The PrefixTree class accepts any strings of single-width characters. 
Generally speaking, the PrefixTree is a less efficient data structure than the RadixTree. This class is provided for comparison and completeness. 

\code{$print()} and \code{$to_string()} prints to screen or outputs the tree to a string representation.

\code{$to_dataframe()} outputs all sequences held by the tree. 

\code{$size()} outputs the size of the tree (i.e. how many sequences are contained). 

\code{$insert()}, \code{$erase()} and \code{$find()} insert, erase and find sequences in the tree, respectively.

\code{$search()} This function searches for similar sequences within a threshold (given by max_distance or max_fraction) based on Levenshtein or Hamming algorithms. 
The output of this function is a data.frame of all matches with columns "query" (the sequences input to the search function), 
"target" (the sequences inserted into the tree) and "distance" the absolute distance between query and target sequences.
}

\examples{
# Plot Data: x,y,r
tree <- PrefixTree$new()
tree$insert(c("ACGT", "AAAA"))
tree$erase("AAAA")
tree$search("ACG", max_distance = 1, mode = "levenshtein")
 #   query target distance
 # 1   ACG   ACGT        1
 
tree$search("ACG", max_distance = 1, mode = "hamming")
query    target   distance
<0 rows> (or 0-length row.names)
}
\seealso{
https://en.wikipedia.org/wiki/Trie
}
