% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RadixS7.R
\name{single_gap_search}
\alias{single_gap_search}
\title{Single-gap alignment search}
\usage{
single_gap_search(
  x,
  query,
  max_distance,
  gap_cost = 1L,
  nthreads = 1L,
  show_progress = FALSE
)
}
\arguments{
\item{x}{A \code{radix_tree} object.}

\item{query}{A character vector of query sequences.}

\item{max_distance}{How far to search in units of absolute distance. Can be a single value or a vector. Mutually exclusive with max_fraction.}

\item{gap_cost}{The cost of a gap for use with the "global" or "anchored" distance metrics. See details.}

\item{nthreads}{The number of threads to use for parallel computation.}

\item{show_progress}{Whether to show a progress bar.}
}
\value{
A data frame with columns \code{query}, \code{target}, and \code{distance}.
}
\description{
Single-gap alignment search
}
