% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zz_help_files.R
\name{td_levenshtein}
\alias{td_levenshtein}
\title{td_levenshtein}
\usage{
td_levenshtein(query, subject=NULL, max_distance = NA_integer_, symmetric = FALSE, nthreads = 1L)
}
\arguments{
\item{query}{A character vector of strings as query}

\item{subject}{A character vector of strings as subject, a prefix_tree or NULL. See details}

\item{max_distance}{The max levenshtein distance to search for matches.}

\item{symmetric}{Whether to treat query and subject as the same (i.e. pairwise matching)}

\item{nthreads}{Number of threads to use in computation}
}
\value{
A data.frame containing query, subject and distance.
}
\description{
Performs levenshtein matching using a prefix tree
}
\details{
Subject parameter can be a character vector, a prefix_tree or NULL. If NULL, the function will perform a pairwise search on query.
}
\examples{
x <- td_levenshtein(query = "helo", subject = "hello")
}
