% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zz_help_files.R
\name{td_partial_hamming}
\alias{td_partial_hamming}
\title{td_partial_hamming}
\usage{
td_partial_hamming(query, subject=NULL, max_distance = NA_integer_, symmetric = FALSE, nthreads = 1L)
}
\arguments{
\item{query}{A character vector of strings as query}

\item{subject}{A character vector of strings as subject, a suffix_tree or NULL. See details}

\item{max_distance}{The max hamming distance to search for matches.}

\item{symmetric}{Whether to treat query and subject as the same (i.e. pairwise matching)}

\item{nthreads}{Number of threads to use in computation}
}
\value{
A data.frame containing query, subject and distance.
}
\description{
Performs partial hamming matching using a suffix tree
}
\details{
Subject parameter can be a character vector, a suffix_tree or NULL. If NULL, the function will perform a pairwise search on query.
}
\examples{
x <- td_partial_hamming(query = "ell", subject = "hello")
}
