% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zz_help_files.R
\name{td_prefix_tree}
\alias{td_prefix_tree}
\title{td_prefix_tree}
\usage{
td_prefix_tree(subject)
}
\arguments{
\item{subject}{A character vector}
}
\value{
A prefix tree: a list of class "prefix_tree" containing an external pointer
}
\description{
Creates a prefix tree
}
\details{
A prefix tree can be used to perform levenshtein and hamming calculations.
}
\examples{
x <- td_prefix_tree("hello")
}
