% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zz_help_files.R
\name{td_suffix_tree}
\alias{td_suffix_tree}
\title{td_suffix_tree}
\usage{
td_suffix_tree(subject, min_length)
}
\arguments{
\item{subject}{A character vector}

\item{min_length}{Minimum suffix length in the tree}
}
\value{
A suffix tree: a list of class "suffix_tree" containing an external pointer
}
\description{
Creates a suffix tree
}
\details{
A suffix tree can be used to perform partial hamming matching.
}
\examples{
x <- td_suffix_tree("hello")
}
