% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RadixS7.R
\name{to_string}
\alias{to_string}
\title{Convert a trie or forest to a string}
\usage{
to_string(x)
}
\arguments{
\item{x}{A seqtrie S7 object.}
}
\value{
A string representation of the trie or forest.
}
\description{
Convert a trie or forest to a string
}
