% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RadixS7.R
\name{to_vector}
\alias{to_vector}
\title{Convert sequences to a character vector}
\usage{
to_vector(x)
}
\arguments{
\item{x}{A seqtrie S7 object.}
}
\value{
A character vector of stored sequences.
}
\description{
Convert sequences to a character vector
}
