% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynamic_prog.R
\name{dist_matrix}
\alias{dist_matrix}
\title{Compute distances between all combinations of two sets of sequences}
\usage{
dist_matrix(
  query,
  target,
  mode,
  cost_matrix = NULL,
  gap_cost = NULL,
  nthreads = 1,
  show_progress = FALSE
)
}
\arguments{
\item{query}{A character vector of query sequences}

\item{target}{A character vector of target sequences}

\item{mode}{The distance metric to use. One of hamming (hm), levenshtein (lv) or anchored (an)}

\item{cost_matrix}{A custom cost matrix for use with the "levenshtein" or "anchored" distance metrics}

\item{gap_cost}{The cost of a gap for use with the "levenshtein" or "anchored" distance metrics}

\item{nthreads}{The number of threads to use for parallel computation}

\item{show_progress}{Whether to show a progress bar}
}
\value{
The output of this function is a distance matrix between all query (rows) and target (columns) sequences.
For anchored searches, the output also includes attributes "query_size" and "target_size" which are matrices containing the lengths of the query and target sequences that are aligned.
}
\description{
Compute distances between all combinations of query and target sequences
}
\examples{
dist_matrix(c("ACGT", "AAAA"), c("ACG", "ACGT"), mode = "levenshtein")
}
