% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairwise.R
\name{dist_pairwise}
\alias{dist_pairwise}
\title{Pairwise distance between two sets of sequences}
\usage{
dist_pairwise(
  query,
  target,
  mode,
  cost_matrix = NULL,
  gap_cost = NULL,
  gap_open_cost = NULL,
  nthreads = 1,
  show_progress = FALSE
)
}
\arguments{
\item{query}{`r rdoc("query")`}

\item{target}{`r rdoc("target")`. Must be the same length as query.}

\item{mode}{`r rdoc("mode")`}

\item{cost_matrix}{`r rdoc("cost_matrix")`}

\item{gap_cost}{`r rdoc("gap_cost")`}

\item{gap_open_cost}{`r rdoc("gap_open_cost")`}

\item{nthreads}{`r rdoc("nthreads")`}

\item{show_progress}{`r rdoc("show_progress")`}
}
\value{
The output of this function is a vector of distances. If mode == "anchored" then the output also includes attributes "query_size" and 
"target_size" which are vectors containing the lengths of the query and target sequences that are aligned.
}
\description{
Compute the pairwise distance between two sets of sequences
}
\details{
This function calculates pairwise distances based on Hamming, Levenshtein or Anchored algorithms. _query_ and _target_ must be the same length.
`r rdoc("details")`
}
\examples{
dist_pairwise(c("ACGT", "AAAA"), c("ACG", "ACGT"), mode = "global")
}
